refactor: split voting files
